fix(server): Fix cursor position offset; Add tests for array quotes

+++ server/doc-values/value-key-enum-assignment.go
@@ -209,7 +209,7 @@ func (v KeyEnumAssignmentValue) FetchCompletions(value string, cursor common.Cur
 			return nil
 		}
 
-		start := uint32(foundPosition + len(v.Separator))
+		start := foundPosition + len(v.Separator)
 		remainingValue := value[start:]
 		remainingCursor := cursor.ShiftHorizontal(-start)
